+++ src/Administration/Resources/app/administration/test/e2e/cypress/integration/settings/sw-settings-country/edit-vat-id-field-required.spec.js
@@ -12,13 +12,17 @@ describe('Country: Test can setting VAT id field required', () => {
     });
 
     it('@settings: can setting VAT id field required', () => {
-        cy.onlyOnFeature('FEATURE_NEXT_14114');
         cy.openInitialPage(`${Cypress.env('admin')}#/sw/settings/login/registration/index`);
         cy.server();
         cy.route({
             url: `${Cypress.env('apiPath')}/_action/system-config/batch`,
             method: 'post'
         }).as('saveSettings');
+
+        cy.route({
+            url: `${Cypress.env('apiPath')}/country/*`,
+            method: 'patch'
+        }).as('saveCountry');
 
         cy.get('input[name="core.loginRegistration.showAccountTypeSelection"]').scrollIntoView();
         cy.get('input[name="core.loginRegistration.showAccountTypeSelection"]').should('be.visible');
@@ -30,7 +34,7 @@ describe('Country: Test can setting VAT id field required', () => {
             expect(xhr).to.have.property('status', 204);
         });
 
-        cy.openInitialPage(`${Cypress.env('admin')}#/sw/settings/country/index`);
+        cy.visit(`${Cypress.env('admin')}#/sw/settings/country/index`);
         const settingPage = new SettingsPageObject();
 
         cy.get('.sw-admin-menu__item--sw-settings').click();
@@ -39,6 +43,10 @@ describe('Country: Test can setting VAT id field required', () => {
         cy.get(`${settingPage.elements.dataGridRow}--0 ${settingPage.elements.countryColumnName}`).click();
         cy.get('.sw-settings-country-general__vat-id-required .sw-field--switch__input').click();
         cy.get(settingPage.elements.countrySaveAction).click();
+
+        cy.wait('@saveCountry').then((xhr) => {
+            expect(xhr).to.have.property('status', 204);
+        });
 
         cy.visit('/account/login');
         const accountTypeSelector = '.register-form select[name="accountType"]';