+++ src/Administration/Resources/app/administration/test/module/sw-settings-country/sw-settings-country-detail.spec.js
@@ -6,7 +6,7 @@ import 'src/app/component/base/sw-container';
 import 'src/app/component/base/sw-tabs';
 import 'src/app/component/base/sw-tabs-item';
 
-function createWrapper(privileges = [], isActiveFeature) {
+function createWrapper(privileges = []) {
     const localVue = createLocalVue();
     localVue.directive('tooltip', {});
 
@@ -81,9 +81,6 @@ function createWrapper(privileges = [], isActiveFeature) {
                     return privileges.includes(identifier);
                 }
             },
-            feature: {
-                isActive: () => isActiveFeature
-            },
             customFieldDataProviderService: {
                 getCustomFieldSets: () => Promise.resolve([])
             }
@@ -157,171 +154,10 @@ describe('module/sw-settings-country/page/sw-settings-country-detail', () => {
         expect(wrapper.vm).toBeTruthy();
     });
 
-    it('should be able to save the country', async () => {
-        const wrapper = createWrapper([
-            'country.editor'
-        ], false);
-        await wrapper.vm.$nextTick();
-
-        const saveButton = wrapper.find(
-            '.sw-settings-country-detail__save-action'
-        );
-        const countryNameField = wrapper.find(
-            'sw-field-stub[label="sw-settings-country.detail.labelName"]'
-        );
-        const countryPositionField = wrapper.find(
-            'sw-field-stub[label="sw-settings-country.detail.labelPosition"]'
-        );
-        const countryIsoField = wrapper.find(
-            'sw-field-stub[label="sw-settings-country.detail.labelIso"]'
-        );
-        const countryIso3Field = wrapper.find(
-            'sw-field-stub[label="sw-settings-country.detail.labelIso3"]'
-        );
-        const countryActiveField = wrapper.find(
-            'sw-field-stub[label="sw-settings-country.detail.labelActive"]'
-        );
-        const countryShippingAvailableField = wrapper.find(
-            'sw-field-stub[label="sw-settings-country.detail.labelShippingAvailable"]'
-        );
-        const countryTaxFreeField = wrapper.find(
-            'sw-field-stub[label="sw-settings-country.detail.labelTaxFree"]'
-        );
-        const countryCompaniesTaxFreeField = wrapper.find(
-            'sw-switch-field-stub[label="sw-settings-country.detail.labelCompanyTaxFree"]'
-        );
-        const countryCheckVatIdFormatField = wrapper.find(
-            'sw-switch-field-stub[label="sw-settings-country.detail.labelCheckVatIdFormat"]'
-        );
-        const countryForceStateInRegistrationField = wrapper.find(
-            'sw-field-stub[label="sw-settings-country.detail.labelForceStateInRegistration"]'
-        );
-
-        expect(saveButton.attributes().disabled).toBeFalsy();
-        expect(countryNameField.attributes().disabled).toBeUndefined();
-        expect(countryPositionField.attributes().disabled).toBeUndefined();
-        expect(countryIsoField.attributes().disabled).toBeUndefined();
-        expect(countryIso3Field.attributes().disabled).toBeUndefined();
-        expect(countryActiveField.attributes().disabled).toBeUndefined();
-        expect(countryShippingAvailableField.attributes().disabled).toBeUndefined();
-        expect(countryTaxFreeField.attributes().disabled).toBeUndefined();
-        expect(countryCompaniesTaxFreeField.attributes().disabled).toBeUndefined();
-        expect(countryCheckVatIdFormatField.attributes().disabled).toBeUndefined();
-        expect(countryForceStateInRegistrationField.attributes().disabled).toBeUndefined();
-    });
-
-    it('should not be able to save the country', async () => {
-        const wrapper = createWrapper([], false);
-
-        await wrapper.vm.$nextTick();
-
-        const saveButton = wrapper.find(
-            '.sw-settings-country-detail__save-action'
-        );
-        const countryNameField = wrapper.find(
-            'sw-field-stub[label="sw-settings-country.detail.labelName"]'
-        );
-        const countryPositionField = wrapper.find(
-            'sw-field-stub[label="sw-settings-country.detail.labelPosition"]'
-        );
-        const countryIsoField = wrapper.find(
-            'sw-field-stub[label="sw-settings-country.detail.labelIso"]'
-        );
-        const countryIso3Field = wrapper.find(
-            'sw-field-stub[label="sw-settings-country.detail.labelIso3"]'
-        );
-        const countryActiveField = wrapper.find(
-            'sw-field-stub[label="sw-settings-country.detail.labelActive"]'
-        );
-        const countryShippingAvailableField = wrapper.find(
-            'sw-field-stub[label="sw-settings-country.detail.labelShippingAvailable"]'
-        );
-        const countryTaxFreeField = wrapper.find(
-            'sw-field-stub[label="sw-settings-country.detail.labelTaxFree"]'
-        );
-        const countryCompaniesTaxFreeField = wrapper.find(
-            'sw-switch-field-stub[label="sw-settings-country.detail.labelCompanyTaxFree"]'
-        );
-        const countryCheckVatIdFormatField = wrapper.find(
-            'sw-switch-field-stub[label="sw-settings-country.detail.labelCheckVatIdFormat"]'
-        );
-        const countryForceStateInRegistrationField = wrapper.find(
-            'sw-field-stub[label="sw-settings-country.detail.labelForceStateInRegistration"]'
-        );
-
-        expect(saveButton.attributes().disabled).toBeTruthy();
-        expect(countryNameField.attributes().disabled).toBeTruthy();
-        expect(countryPositionField.attributes().disabled).toBeTruthy();
-        expect(countryIsoField.attributes().disabled).toBeTruthy();
-        expect(countryIso3Field.attributes().disabled).toBeTruthy();
-        expect(countryActiveField.attributes().disabled).toBeTruthy();
-        expect(countryShippingAvailableField.attributes().disabled).toBeTruthy();
-        expect(countryTaxFreeField.attributes().disabled).toBeTruthy();
-        expect(countryCompaniesTaxFreeField.attributes().disabled).toBeTruthy();
-        expect(countryCheckVatIdFormatField.attributes().disabled).toBeTruthy();
-        expect(countryForceStateInRegistrationField.attributes().disabled).toBeTruthy();
-    });
-
-    it('should be able to create a new country state', async () => {
-        const wrapper = createWrapper([
-            'country.editor'
-        ], false);
-        await wrapper.vm.$nextTick();
-
-        const createButton = wrapper.find('.sw-settings-country-detail__add-country-state-button');
-
-        expect(createButton.attributes().disabled).toBeFalsy();
-    });
-
-    it('should not be able to create a new country state', async () => {
-        const wrapper = createWrapper([], false);
-        await wrapper.vm.$nextTick();
-
-        const createButton = wrapper.find('.sw-settings-country-detail__add-country-state-button');
-
-        expect(createButton.attributes().disabled).toBeTruthy();
-    });
-
-    it('should be able to edit a country state', async () => {
-        const wrapper = createWrapper([
-            'country.editor'
-        ], false);
-        await wrapper.vm.$nextTick();
-
-        const editMenuItem = wrapper.find('.sw-settings-country-detail__edit-country-state-action');
-        expect(editMenuItem.attributes().disabled).toBeFalsy();
-    });
-
-    it('should not be able to edit a country state', async () => {
-        const wrapper = createWrapper([], false);
-        await wrapper.vm.$nextTick();
-
-        const editMenuItem = wrapper.find('.sw-settings-country-detail__edit-country-state-action');
-        expect(editMenuItem.attributes().disabled).toBeTruthy();
-    });
-
-    it('should be able to delete a country state', async () => {
-        const wrapper = createWrapper([
-            'country.editor'
-        ], false);
-        await wrapper.vm.$nextTick();
-
-        const editMenuItem = wrapper.find('.sw-one-to-many-grid__delete-action');
-        expect(editMenuItem.attributes().disabled).toBeFalsy();
-    });
-
-    it('should not be able to delete a country state', async () => {
-        const wrapper = createWrapper([], false);
-        await wrapper.vm.$nextTick();
-
-        const editMenuItem = wrapper.find('.sw-one-to-many-grid__delete-action');
-        expect(editMenuItem.attributes().disabled).toBeTruthy();
-    });
-
     it('should be render tab', async () => {
         const wrapper = createWrapper([
             'country.editor'
-        ], true);
+        ]);
 
         await wrapper.vm.$nextTick();
         const generalTab = wrapper.find('.sw-settings-country__setting-tab');
@@ -334,7 +170,7 @@ describe('module/sw-settings-country/page/sw-settings-country-detail', () => {
     it('should be able to save the country', async () => {
         const wrapper = createWrapper([
             'country.editor'
-        ], true);
+        ]);
         await wrapper.vm.$nextTick();
 
         const saveButton = wrapper.find(
@@ -345,7 +181,7 @@ describe('module/sw-settings-country/page/sw-settings-country-detail', () => {
     });
 
     it('should not be able to save the country', async () => {
-        const wrapper = createWrapper([], true);
+        const wrapper = createWrapper([]);
         await wrapper.vm.$nextTick();
 
         const saveButton = wrapper.find(