+++ src/Core/Checkout/Test/Document/DocumentType/InvoiceServiceTest.php
@@ -26,7 +26,6 @@ use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\DataAbstractionLayer\Exception\InconsistentCriteriaIdsException;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\DataAbstractionLayer\TaxFreeConfig;
-use Shopware\Core\Framework\Feature;
 use Shopware\Core\Framework\Rule\Collector\RuleConditionRegistry;
 use Shopware\Core\Framework\Test\TestCaseBase\CountryAddToSalesChannelTestBehaviour;
 use Shopware\Core\Framework\Test\TestCaseBase\IntegrationTestBehaviour;
@@ -182,11 +181,7 @@ class InvoiceServiceTest extends TestCase
         /** @var OrderEntity $order */
         $order = $this->getOrderById($orderId);
         $country = $order->getDeliveries()->getShippingAddress()->getCountries()->first();
-        if (Feature::isActive('FEATURE_NEXT_14114')) {
-            $country->setCompanyTax(new TaxFreeConfig(true, Defaults::CURRENCY, 0));
-        } else {
-            $country->setCompanyTaxFree(true);
-        }
+        $country->setCompanyTax(new TaxFreeConfig(true, Defaults::CURRENCY, 0));
         $companyPhone = '123123123';
         $vatIds = ['VAT-123123'];
         $order->getOrderCustomer()->setVatIds($vatIds);