+++ src/Core/System/Currency/CurrencyEntity.php
@@ -111,8 +111,6 @@ class CurrencyEntity extends Entity
     protected $totalRounding;
 
     /**
-     * @internal (flag:FEATURE_NEXT_14114)
-     *
      * @var float|null
      */
     protected $taxFreeFrom;
@@ -305,17 +303,11 @@ class CurrencyEntity extends Entity
         return $this->itemRounding->getDecimals();
     }
 
-    /**
-     * @internal (flag:FEATURE_NEXT_14114)
-     */
     public function getTaxFreeFrom(): ?float
     {
         return $this->taxFreeFrom;
     }
 
-    /**
-     * @internal (flag:FEATURE_NEXT_14114)
-     */
     public function setTaxFreeFrom(?float $taxFreeFrom): void
     {
         $this->taxFreeFrom = $taxFreeFrom;